split messaging internal tests
