+++ code/messaging/src/test/java/com/adobe/marketing/mobile/MessagingInternalTests.java
@@ -12,39 +12,32 @@
 
 package com.adobe.marketing.mobile;
 
-import android.app.Activity;
 import android.app.Application;
 import android.content.Context;
 import android.content.pm.ApplicationInfo;
 import android.content.pm.PackageManager;
 import android.os.Bundle;
 
-import org.json.JSONObject;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
-import org.mockito.Mockito;
 import org.powermock.api.mockito.PowerMockito;
 import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
 import org.powermock.reflect.Whitebox;
 
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.ExecutorService;
 
-import static com.adobe.marketing.mobile.MessagingConstants.EventDataKeys.Messaging.REFRESH_MESSAGES;
 import static org.junit.Assert.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyInt;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.ArgumentMatchers.nullable;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
@@ -55,10 +48,7 @@ import static org.mockito.Mockito.when;
 public class MessagingInternalTests {
 
     private MessagingInternal messagingInternal;
-    private AndroidPlatformServices platformServices;
-    private JsonUtilityService jsonUtilityService;
     private EventHub eventHub;
-    private Map<String, Object> mockConfigState = new HashMap<>();
 
     // Mocks
     @Mock
@@ -70,8 +60,6 @@ public class MessagingInternalTests {
     @Mock
     Map<String, Object> mockEdgeIdentityData;
     @Mock
-    EventData mockEdgeIdentityEventData;
-    @Mock
     ConcurrentLinkedQueue<Event> mockEventQueue;
     @Mock
     Application mockApplication;
@@ -80,13 +68,7 @@ public class MessagingInternalTests {
     @Mock
     Core mockCore;
     @Mock
-    Activity mockActivity;
-    @Mock
-    MessagingFullscreenMessage mockFullscreenMessage;
-    @Mock
     AndroidPlatformServices mockPlatformServices;
-    @Mock
-    UIService mockUIService;
     @Mock
     PackageManager packageManager;
     @Mock
@@ -99,28 +81,21 @@ public class MessagingInternalTests {
         PowerMockito.mockStatic(MobileCore.class);
         PowerMockito.mockStatic(Event.class);
         PowerMockito.mockStatic(App.class);
-        platformServices = new AndroidPlatformServices();
-        jsonUtilityService = platformServices.getJsonUtilityService();
         eventHub = new EventHub("testEventHub", mockPlatformServices);
         mockCore.eventHub = eventHub;
         when(MobileCore.getCore()).thenReturn(mockCore);
-        when(App.getCurrentActivity()).thenReturn(mockActivity);
-        when(mockApplication.getApplicationContext()).thenReturn(context);
-        when(context.getPackageName()).thenReturn("mock_placement");
-        // mocks for getting activity id from manifest
+
+        // activity id mocks
         when(App.getApplication()).thenReturn(mockApplication);
         when(mockApplication.getPackageManager()).thenReturn(packageManager);
+        when(mockApplication.getApplicationContext()).thenReturn(context);
         when(packageManager.getApplicationInfo(anyString(), anyInt())).thenReturn(applicationInfo);
         Whitebox.setInternalState(applicationInfo, "metaData", bundle);
         when(bundle.getString(anyString())).thenReturn("mock_activity");
-
+        // placement id mocks
         when(mockApplication.getPackageName()).thenReturn("mock_placement");
-        when(mockPlatformServices.getJsonUtilityService()).thenReturn(jsonUtilityService);
-        when(mockPlatformServices.getUIService()).thenReturn(mockUIService);
-        when(mockExtensionApi.getSharedEventState(anyString(), any(Event.class), nullable(ExtensionErrorCallback.class))).thenReturn(mockConfigState);
+
         messagingInternal = new MessagingInternal(mockExtensionApi);
-        Whitebox.setInternalState(messagingInternal, "messagingState", messagingState);
-        mockConfigState.put(MessagingConstants.SharedState.Configuration.ORG_ID, "mock_activity");
     }
 
     // ========================================================================================
@@ -663,1129 +638,5 @@ public class MessagingInternalTests {
 
         // verify
         assertEquals("Gets the same executor instance on the next get", executorService, messagingInternal.getExecutor());
-    }
-
-    // ========================================================================================
-    // IAM rules retrieval from Offers
-    // ========================================================================================
-    @Test
-    public void test_fetchMessages_CalledOnExtensionStart() {
-        // setup
-        final ArgumentCaptor<Event> eventCaptor = ArgumentCaptor.forClass(Event.class);
-        // expected dispatched event data
-        String expectedOffersEventData = "{\"decisionscopes\":[{\"activityId\":\"mock_activity\",\"placementId\":\"mock_placement\",\"itemCount\":30}],\"type\":\"prefetch\"}";
-
-        // verify dispatch event is called
-        // 1 event dispatched: Offers iam fetch event when extension is registered
-        PowerMockito.verifyStatic(MobileCore.class, times(1));
-        MobileCore.dispatchEvent(eventCaptor.capture(), any(ExtensionErrorCallback.class));
-
-        // verify events
-        Event event = eventCaptor.getAllValues().get(0);
-        assertNotNull(event.getData());
-        assertEquals(expectedOffersEventData, event.getData().toString());
-    }
-
-    @Test
-    public void test_refreshInAppMessages_Invoked() {
-        // setup
-        final ArgumentCaptor<Event> eventCaptor = ArgumentCaptor.forClass(Event.class);
-        // trigger event
-        EventData eventData = new EventData();
-        eventData.putBoolean(REFRESH_MESSAGES, true);
-        // expected dispatched event data
-        String expectedOffersEventData = "{\"decisionscopes\":[{\"activityId\":\"mock_activity\",\"placementId\":\"mock_placement\",\"itemCount\":30}],\"type\":\"prefetch\"}";
-
-        // Mocks
-        Event mockEvent = mock(Event.class);
-        // when mock event getType called return MESSAGING
-        when(mockEvent.getType()).thenReturn(MessagingConstants.EventType.MESSAGING);
-
-        // when mock event getSource called return REQUEST_CONTENT
-        when(mockEvent.getSource()).thenReturn(EventSource.REQUEST_CONTENT.getName());
-
-        // when get eventData called return data with "REFRESH_MESSAGES, true"
-        when(mockEvent.getEventData()).thenReturn(eventData.toObjectMap());
-
-        // test
-        messagingInternal.queueEvent(mockEvent);
-        messagingInternal.processEvents();
-
-        // verify dispatch event is called
-        // 2 events dispatched: Offers iam fetch event when extension is registered + Offers iam fetch event when refresh in app messages event is received
-        PowerMockito.verifyStatic(MobileCore.class, times(2));
-        MobileCore.dispatchEvent(eventCaptor.capture(), any(ExtensionErrorCallback.class));
-
-        // verify events
-        Event event = eventCaptor.getAllValues().get(0);
-        assertNotNull(event.getData());
-        assertEquals(expectedOffersEventData, event.getData().toString());
-        event = eventCaptor.getAllValues().get(1);
-        assertNotNull(event.getData());
-        assertEquals(expectedOffersEventData, event.getData().toString());
-    }
-
-    // ========================================================================================
-    // Offers rules payload processing
-    // ========================================================================================
-    @Test
-    public void test_handleEdgeResponseEvent_ValidOffersIAMPayloadPresent() throws Exception {
-        // setup
-        // trigger event
-        HashMap<String, Object> eventData = new HashMap<>();
-        eventData.put("type", "personalization:decisions");
-        eventData.put("requestEventId", "2E964037-E319-4D14-98B8-0682374E547B");
-        JSONObject payload = new JSONObject("    {\n" +
-                "      \"activity\" : {\n" +
-                "        \"id\" : \"mock_activity\",\n" +
-                "        \"etag\" : \"2\"\n" +
-                "      },\n" +
-                "      \"scope\" : \"eyJhY3Rpdml0eUlkIjoieGNvcmU6b2ZmZXItYWN0aXZpdHk6MTMyM2RiZTk0ZjJlZWY5MyIsInBsYWNlbWVudElkIjoieGNvcmU6b2ZmZXItcGxhY2VtZW50OjEzMjNkOWViNDNhYWNhZGEiLCJpdGVtQ291bnQiOjMwfQ==\",\n" +
-                "      \"placement\" : {\n" +
-                "        \"id\" : \"mock_placement\",\n" +
-                "        \"etag\" : \"1\"\n" +
-                "      },\n" +
-                "      \"items\" : [\n" +
-                "        {\n" +
-                "          \"id\" : \"xcore:fallback-offer:1323dbbc1c6eef91\",\n" +
-                "          \"data\" : {\n" +
-                "            \"id\" : \"xcore:fallback-offer:1323dbbc1c6eef91\",\n" +
-                "            \"format\" : \"application\\/json\",\n" +
-                "            \"content\" : \"{ \\\"version\\\": 1, \\\"rules\\\": [ { \\\"condition\\\": { \\\"type\\\": \\\"group\\\", \\\"definition\\\": { \\\"logic\\\": \\\"and\\\", \\\"conditions\\\": [ { \\\"definition\\\": { \\\"key\\\": \\\"contextdata.testShowMessage\\\", \\\"matcher\\\": \\\"eq\\\", \\\"values\\\": [ \\\"true\\\" ] }, \\\"type\\\": \\\"matcher\\\" } ] } }, \\\"consequences\\\": [ { \\\"id\\\": \\\"341800180\\\", \\\"type\\\": \\\"cjmiam\\\", \\\"detail\\\": { \\\"remoteAssets\\\": [], \\\"html\\\": \\\"<html><head><\\/head><body bgcolor=\\\\\\\"black\\\\\\\"><br \\/><br \\/><br \\/><br \\/><br \\/><br \\/><h1 align=\\\\\\\"center\\\\\\\" style=\\\\\\\"color: white;\\\\\\\">IN-APP MESSAGING POWERED BY <br \\/>OFFER DECISIONING<\\/h1><h1 align=\\\\\\\"center\\\\\\\"><a style=\\\\\\\"color: white;\\\\\\\" href=\\\\\\\"adbinapp:\\/\\/cancel\\\\\\\" >dismiss me<\\/a><\\/h1><\\/body><\\/html>\\\", \\\"template\\\": \\\"fullscreen\\\" } } ] } ] }\"\n" +
-                "          },\n" +
-                "          \"etag\" : \"1\",\n" +
-                "          \"schema\" : \"https:\\/\\/ns.adobe.com\\/experience\\/offer-management\\/content-component-json\"\n" +
-                "        }\n" +
-                "      ],\n" +
-                "      \"id\" : \"cb25ecb0-d085-44ac-b73d-797a3265d37c\"\n" +
-                "    }\n" +
-                "  ]");
-        eventData.put("payload", payload);
-        eventData.put("requestId", "D158979E-0506-4968-8031-17A6A8A87DA8");
-        // expected messaging consequence payload
-        String expectedIAMPayload = "{\n" +
-                "        \"triggeredconsequence\" : {\n" +
-                "            \"id\" : \"341800180\",\n" +
-                "            \"detail\" : {\n" +
-                "                \"template\" : \"fullscreen\",\n" +
-                "                \"html\" : \"<html><head></head><body bgcolor=\"black\"><br /><br /><br /><br /><br /><br /><h1 align=\"center\" style=\"color: white;\">IN-APP MESSAGING POWERED BY <br />OFFER DECISIONING</h1><h1 align=\"center\"><a style=\"color: white;\" href=\"adbinapp://cancel\" >dismiss me</a></h1></body></html>\",\n" +
-                "                \"remoteAssets\" : [ ]\n" +
-                "            },\n" +
-                "            \"type\" : \"cjmiam\"\n" +
-                "        }\n" +
-                "    }";
-
-        // Mocks
-        Event mockEvent = mock(Event.class);
-
-        // when mock event getType called return EDGE
-        when(mockEvent.getType()).thenReturn(MessagingConstants.EventType.EDGE);
-
-        // when mock event getSource called return PERSONALIZATION_DECISIONS
-        when(mockEvent.getSource()).thenReturn(MessagingConstants.EventSource.PERSONALIZATION_DECISIONS);
-
-        // when get eventData called return event data containing a valid offers iam payload
-        when(mockEvent.getEventData()).thenReturn(eventData);
-
-        // test
-        messagingInternal.queueEvent(mockEvent);
-        messagingInternal.processEvents();
-
-        // verify rule loaded
-        ConcurrentHashMap loadedRules = mockCore.eventHub.getModuleRuleAssociation();
-        Map.Entry<Module, ConcurrentLinkedQueue<Rule>> entry = (Map.Entry) loadedRules.entrySet().iterator().next();
-        Rule loadedRule = entry.getValue().remove();
-        assertTrue(loadedRule.toString().contains(expectedIAMPayload));
-    }
-
-    @Test
-    public void test_handleEdgeResponseEvent_MultipleValidOffersIAMPayloadPresent() throws Exception {
-        // setup
-        // trigger event
-        HashMap<String, Object> eventData = new HashMap<>();
-        eventData.put("type", "personalization:decisions");
-        eventData.put("requestEventId", "2E964037-E319-4D14-98B8-0682374E547B");
-        JSONObject payload = new JSONObject("    {\n" +
-                "      \"activity\" : {\n" +
-                "        \"id\" : \"mock_activity\",\n" +
-                "        \"etag\" : \"2\"\n" +
-                "      },\n" +
-                "      \"scope\" : \"eyJhY3Rpdml0eUlkIjoieGNvcmU6b2ZmZXItYWN0aXZpdHk6MTMyM2RiZTk0ZjJlZWY5MyIsInBsYWNlbWVudElkIjoieGNvcmU6b2ZmZXItcGxhY2VtZW50OjEzMjNkOWViNDNhYWNhZGEiLCJpdGVtQ291bnQiOjMwfQ==\",\n" +
-                "      \"placement\" : {\n" +
-                "        \"id\" : \"mock_placement\",\n" +
-                "        \"etag\" : \"1\"\n" +
-                "      },\n" +
-                "      \"items\" : [\n" +
-                "        {\n" +
-                "          \"id\" : \"xcore:fallback-offer:1323dbbc1c6eef91\",\n" +
-                "          \"data\" : {\n" +
-                "            \"id\" : \"xcore:fallback-offer:1323dbbc1c6eef91\",\n" +
-                "            \"format\" : \"application\\/json\",\n" +
-                "            \"content\" : \"{ \\\"version\\\": 1, \\\"rules\\\": [ { \\\"condition\\\": { \\\"type\\\": \\\"group\\\", \\\"definition\\\": { \\\"logic\\\": \\\"and\\\", \\\"conditions\\\": [ { \\\"definition\\\": { \\\"key\\\": \\\"contextdata.testShowMessage\\\", \\\"matcher\\\": \\\"eq\\\", \\\"values\\\": [ \\\"true\\\" ] }, \\\"type\\\": \\\"matcher\\\" } ] } }, \\\"consequences\\\": [ { \\\"id\\\": \\\"341800180\\\", \\\"type\\\": \\\"cjmiam\\\", \\\"detail\\\": { \\\"remoteAssets\\\": [], \\\"html\\\": \\\"<html><head><\\/head><body bgcolor=\\\\\\\"black\\\\\\\"><br \\/><br \\/><br \\/><br \\/><br \\/><br \\/><h1 align=\\\\\\\"center\\\\\\\" style=\\\\\\\"color: white;\\\\\\\">IN-APP MESSAGING POWERED BY <br \\/>OFFER DECISIONING<\\/h1><h1 align=\\\\\\\"center\\\\\\\"><a style=\\\\\\\"color: white;\\\\\\\" href=\\\\\\\"adbinapp:\\/\\/cancel\\\\\\\" >dismiss me<\\/a><\\/h1><\\/body><\\/html>\\\", \\\"template\\\": \\\"fullscreen\\\" } } ] } ] }\"\n" +
-                "          },\n" +
-                "          \"etag\" : \"1\",\n" +
-                "          \"schema\" : \"https:\\/\\/ns.adobe.com\\/experience\\/offer-management\\/content-component-json\"\n" +
-                "        },\n" +
-                "        {\n" +
-                "          \"id\" : \"xcore:fallback-offer:1323dbbc1c6eef91\",\n" +
-                "          \"data\" : {\n" +
-                "            \"id\" : \"xcore:fallback-offer:1323dbbc1c6eef91\",\n" +
-                "            \"format\" : \"application\\/json\",\n" +
-                "            \"content\" : \"{ \\\"version\\\": 1, \\\"rules\\\": [ { \\\"condition\\\": { \\\"type\\\": \\\"group\\\", \\\"definition\\\": { \\\"logic\\\": \\\"and\\\", \\\"conditions\\\": [ { \\\"definition\\\": { \\\"key\\\": \\\"contextdata.testShowMessage2\\\", \\\"matcher\\\": \\\"eq\\\", \\\"values\\\": [ \\\"true\\\" ] }, \\\"type\\\": \\\"matcher\\\" } ] } }, \\\"consequences\\\": [ { \\\"id\\\": \\\"341800180\\\", \\\"type\\\": \\\"cjmiam\\\", \\\"detail\\\": { \\\"remoteAssets\\\": [], \\\"html\\\": \\\"<html><head><\\/head><body bgcolor=\\\\\\\"black\\\\\\\"><br \\/><br \\/><br \\/><br \\/><br \\/><br \\/><h1 align=\\\\\\\"center\\\\\\\" style=\\\\\\\"color: white;\\\\\\\">IN-APP MESSAGING MESSAGE 2 POWERED BY <br \\/>OFFER DECISIONING<\\/h1><h1 align=\\\\\\\"center\\\\\\\"><a style=\\\\\\\"color: white;\\\\\\\" href=\\\\\\\"adbinapp:\\/\\/cancel\\\\\\\" >dismiss me2<\\/a><\\/h1><\\/body><\\/html>\\\", \\\"template\\\": \\\"fullscreen\\\" } } ] } ] }\"\n" +
-                "          },\n" +
-                "          \"etag\" : \"1\",\n" +
-                "          \"schema\" : \"https:\\/\\/ns.adobe.com\\/experience\\/offer-management\\/content-component-json\"\n" +
-                "        }\n" +
-                "      ],\n" +
-                "      \"id\" : \"cb25ecb0-d085-44ac-b73d-797a3265d37c\"\n" +
-                "    }\n" +
-                "  ]");
-        eventData.put("payload", payload);
-        eventData.put("requestId", "D158979E-0506-4968-8031-17A6A8A87DA8");
-        // expected messaging consequence payloads
-        String expectedIAMPayload = "{\n" +
-                "        \"triggeredconsequence\" : {\n" +
-                "            \"id\" : \"341800180\",\n" +
-                "            \"detail\" : {\n" +
-                "                \"template\" : \"fullscreen\",\n" +
-                "                \"html\" : \"<html><head></head><body bgcolor=\"black\"><br /><br /><br /><br /><br /><br /><h1 align=\"center\" style=\"color: white;\">IN-APP MESSAGING POWERED BY <br />OFFER DECISIONING</h1><h1 align=\"center\"><a style=\"color: white;\" href=\"adbinapp://cancel\" >dismiss me</a></h1></body></html>\",\n" +
-                "                \"remoteAssets\" : [ ]\n" +
-                "            },\n" +
-                "            \"type\" : \"cjmiam\"\n" +
-                "        }\n" +
-                "    }";
-
-        String secondExpectedIAMPayload = "{\n" +
-                "        \"triggeredconsequence\" : {\n" +
-                "            \"id\" : \"341800180\",\n" +
-                "            \"detail\" : {\n" +
-                "                \"template\" : \"fullscreen\",\n" +
-                "                \"html\" : \"<html><head></head><body bgcolor=\"black\"><br /><br /><br /><br /><br /><br /><h1 align=\"center\" style=\"color: white;\">IN-APP MESSAGING MESSAGE 2 POWERED BY <br />OFFER DECISIONING</h1><h1 align=\"center\"><a style=\"color: white;\" href=\"adbinapp://cancel\" >dismiss me2</a></h1></body></html>\",\n" +
-                "                \"remoteAssets\" : [ ]\n" +
-                "            },\n" +
-                "            \"type\" : \"cjmiam\"\n" +
-                "        }\n" +
-                "    }";
-
-        // Mocks
-        Event mockEvent = mock(Event.class);
-
-        // when mock event getType called return EDGE
-        when(mockEvent.getType()).thenReturn(MessagingConstants.EventType.EDGE);
-
-        // when mock event getSource called return PERSONALIZATION_DECISIONS
-        when(mockEvent.getSource()).thenReturn(MessagingConstants.EventSource.PERSONALIZATION_DECISIONS);
-
-        // when get eventData called return event data containing multiple valid offers iam payloads
-        when(mockEvent.getEventData()).thenReturn(eventData);
-
-        // test
-        messagingInternal.queueEvent(mockEvent);
-        messagingInternal.processEvents();
-
-        // verify rule loaded
-        ConcurrentHashMap loadedRules = mockCore.eventHub.getModuleRuleAssociation();
-        Map.Entry<Module, ConcurrentLinkedQueue<Rule>> entry = (Map.Entry) loadedRules.entrySet().iterator().next();
-        assertEquals(2, entry.getValue().size());
-        Rule loadedRule = entry.getValue().remove();
-        assertTrue(loadedRule.toString().contains(expectedIAMPayload));
-        entry = (Map.Entry) loadedRules.entrySet().iterator().next();
-        loadedRule = entry.getValue().remove();
-        assertTrue(loadedRule.toString().contains(secondExpectedIAMPayload));
-    }
-
-    @Test
-    public void test_handleEdgeResponseEvent_OneInvalidIAMPayloadPresent() throws Exception {
-        // setup
-        // private mocks
-        Whitebox.setInternalState(messagingInternal, "messagingState", messagingState);
-        // trigger event
-        HashMap<String, Object> eventData = new HashMap<>();
-        eventData.put("type", "personalization:decisions");
-        eventData.put("requestEventId", "2E964037-E319-4D14-98B8-0682374E547B");
-        JSONObject payload = new JSONObject("    {\n" +
-                "      \"activity\" : {\n" +
-                "        \"id\" : \"mock_activity\",\n" +
-                "        \"etag\" : \"2\"\n" +
-                "      },\n" +
-                "      \"scope\" : \"eyJhY3Rpdml0eUlkIjoieGNvcmU6b2ZmZXItYWN0aXZpdHk6MTMyM2RiZTk0ZjJlZWY5MyIsInBsYWNlbWVudElkIjoieGNvcmU6b2ZmZXItcGxhY2VtZW50OjEzMjNkOWViNDNhYWNhZGEiLCJpdGVtQ291bnQiOjMwfQ==\",\n" +
-                "      \"placement\" : {\n" +
-                "        \"id\" : \"mock_placement\",\n" +
-                "        \"etag\" : \"1\"\n" +
-                "      },\n" +
-                "      \"items\" : [\n" +
-                "        {\n" +
-                "          \"id\" : \"xcore:fallback-offer:1323dbbc1c6eef91\",\n" +
-                "          \"data\" : {\n" +
-                "            \"id\" : \"xcore:fallback-offer:1323dbbc1c6eef91\",\n" +
-                "            \"format\" : \"application\\/json\",\n" +
-                "            \"content\" : \"{ \\\"version\\\": 1, \\\"rules\\\": [ { \\\"condition\\\": { \\\"type\\\": \\\"group\\\", \\\"definition\\\": { \\\"logic\\\": \\\"and\\\", \\\"conditions\\\": [ { \\\"definition\\\": { \\\"key\\\": \\\"contextdata.testShowMessage\\\", \\\"matcher\\\": \\\"eq\\\", \\\"values\\\": [ \\\"true\\\" ] }, \\\"type\\\": \\\"matcher\\\" } ] } }, \\\"consequences\\\": [ { \\\"id\\\": \\\"341800180\\\", \\\"detail\\\": { \\\"remoteAssets\\\": [], \\\"html\\\": \\\"<html><head><\\/head><body bgcolor=\\\\\\\"black\\\\\\\"><br \\/><br \\/><br \\/><br \\/><br \\/><br \\/><h1 align=\\\\\\\"center\\\\\\\" style=\\\\\\\"color: white;\\\\\\\">IN-APP MESSAGING POWERED BY <br \\/>OFFER DECISIONING<\\/h1><h1 align=\\\\\\\"center\\\\\\\"><a style=\\\\\\\"color: white;\\\\\\\" href=\\\\\\\"adbinapp:\\/\\/cancel\\\\\\\" >dismiss me<\\/a><\\/h1><\\/body><\\/html>\\\", \\\"template\\\": \\\"fullscreen\\\" } } ] } ] }\"\n" +
-                "          },\n" +
-                "          \"etag\" : \"1\",\n" +
-                "          \"schema\" : \"https:\\/\\/ns.adobe.com\\/experience\\/offer-management\\/content-component-json\"\n" +
-                "        },\n" +
-                "        {\n" +
-                "          \"id\" : \"xcore:fallback-offer:1323dbbc1c6eef91\",\n" +
-                "          \"data\" : {\n" +
-                "            \"id\" : \"xcore:fallback-offer:1323dbbc1c6eef91\",\n" +
-                "            \"format\" : \"application\\/json\",\n" +
-                "            \"content\" : \"{ \\\"version\\\": 1, \\\"rules\\\": [ { \\\"condition\\\": { \\\"type\\\": \\\"group\\\", \\\"definition\\\": { \\\"logic\\\": \\\"and\\\", \\\"conditions\\\": [ { \\\"definition\\\": { \\\"key\\\": \\\"contextdata.testShowMessage2\\\", \\\"matcher\\\": \\\"eq\\\", \\\"values\\\": [ \\\"true\\\" ] }, \\\"type\\\": \\\"matcher\\\" } ] } }, \\\"consequences\\\": [ { \\\"id\\\": \\\"341800180\\\", \\\"type\\\": \\\"cjmiam\\\", \\\"detail\\\": { \\\"remoteAssets\\\": [], \\\"html\\\": \\\"<html><head><\\/head><body bgcolor=\\\\\\\"black\\\\\\\"><br \\/><br \\/><br \\/><br \\/><br \\/><br \\/><h1 align=\\\\\\\"center\\\\\\\" style=\\\\\\\"color: white;\\\\\\\">IN-APP MESSAGING MESSAGE 2 POWERED BY <br \\/>OFFER DECISIONING<\\/h1><h1 align=\\\\\\\"center\\\\\\\"><a style=\\\\\\\"color: white;\\\\\\\" href=\\\\\\\"adbinapp:\\/\\/cancel\\\\\\\" >dismiss me2<\\/a><\\/h1><\\/body><\\/html>\\\", \\\"template\\\": \\\"fullscreen\\\" } } ] } ] }\"\n" +
-                "          },\n" +
-                "          \"etag\" : \"1\",\n" +
-                "          \"schema\" : \"https:\\/\\/ns.adobe.com\\/experience\\/offer-management\\/content-component-json\"\n" +
-                "        }\n" +
-                "      ],\n" +
-                "      \"id\" : \"cb25ecb0-d085-44ac-b73d-797a3265d37c\"\n" +
-                "    }\n" +
-                "  ]");
-        eventData.put("payload", payload);
-        eventData.put("requestId", "D158979E-0506-4968-8031-17A6A8A87DA8");
-        // expected messaging consequence payloads
-        String expectedIAMPayload = "{\n" +
-                "        \"triggeredconsequence\" : {\n" +
-                "            \"id\" : \"341800180\",\n" +
-                "            \"detail\" : {\n" +
-                "                \"template\" : \"fullscreen\",\n" +
-                "                \"html\" : \"<html><head></head><body bgcolor=\"black\"><br /><br /><br /><br /><br /><br /><h1 align=\"center\" style=\"color: white;\">IN-APP MESSAGING MESSAGE 2 POWERED BY <br />OFFER DECISIONING</h1><h1 align=\"center\"><a style=\"color: white;\" href=\"adbinapp://cancel\" >dismiss me2</a></h1></body></html>\",\n" +
-                "                \"remoteAssets\" : [ ]\n" +
-                "            },\n" +
-                "            \"type\" : \"cjmiam\"\n" +
-                "        }\n" +
-                "    }";
-
-        // Mocks
-        Event mockEvent = mock(Event.class);
-
-        // when mock event getType called return EDGE
-        when(mockEvent.getType()).thenReturn(MessagingConstants.EventType.EDGE);
-
-        // when mock event getSource called return PERSONALIZATION_DECISIONS
-        when(mockEvent.getSource()).thenReturn(MessagingConstants.EventSource.PERSONALIZATION_DECISIONS);
-
-        // when get eventData called return event data containing one valid and one invalid offers iam payloads
-        when(mockEvent.getEventData()).thenReturn(eventData);
-
-        // test
-        messagingInternal.queueEvent(mockEvent);
-        messagingInternal.processEvents();
-
-        // verify rule loaded
-        ConcurrentHashMap loadedRules = mockCore.eventHub.getModuleRuleAssociation();
-        Map.Entry<Module, ConcurrentLinkedQueue<Rule>> entry = (Map.Entry) loadedRules.entrySet().iterator().next();
-        assertEquals(1, entry.getValue().size());
-        Rule loadedRule = entry.getValue().remove();
-        assertTrue(loadedRule.toString().contains(expectedIAMPayload));
-    }
-
-    @Test
-    public void test_handleEdgeResponseEvent_OffersIAMPayloadMissingMessageId() throws Exception {
-        // setup
-        // private mocks
-        Whitebox.setInternalState(messagingInternal, "messagingState", messagingState);
-        // trigger event
-        HashMap<String, Object> eventData = new HashMap<>();
-        eventData.put("type", "personalization:decisions");
-        eventData.put("requestEventId", "2E964037-E319-4D14-98B8-0682374E547B");
-        JSONObject payload = new JSONObject("    {\n" +
-                "      \"activity\" : {\n" +
-                "        \"id\" : \"mock_activity\",\n" +
-                "        \"etag\" : \"2\"\n" +
-                "      },\n" +
-                "      \"scope\" : \"eyJhY3Rpdml0eUlkIjoieGNvcmU6b2ZmZXItYWN0aXZpdHk6MTMyM2RiZTk0ZjJlZWY5MyIsInBsYWNlbWVudElkIjoieGNvcmU6b2ZmZXItcGxhY2VtZW50OjEzMjNkOWViNDNhYWNhZGEiLCJpdGVtQ291bnQiOjMwfQ==\",\n" +
-                "      \"placement\" : {\n" +
-                "        \"id\" : \"mock_placement\",\n" +
-                "        \"etag\" : \"1\"\n" +
-                "      },\n" +
-                "      \"items\" : [\n" +
-                "        {\n" +
-                "          \"id\" : \"xcore:fallback-offer:1323dbbc1c6eef91\",\n" +
-                "          \"data\" : {\n" +
-                "            \"id\" : \"xcore:fallback-offer:1323dbbc1c6eef91\",\n" +
-                "            \"format\" : \"application\\/json\",\n" +
-                "            \"content\" : \"{ \\\"version\\\": 1, \\\"rules\\\": [ { \\\"condition\\\": { \\\"type\\\": \\\"group\\\", \\\"definition\\\": { \\\"logic\\\": \\\"and\\\", \\\"conditions\\\": [ { \\\"definition\\\": { \\\"key\\\": \\\"contextdata.testShowMessage\\\", \\\"matcher\\\": \\\"eq\\\", \\\"values\\\": [ \\\"true\\\" ] }, \\\"type\\\": \\\"matcher\\\" } ] } }, \\\"consequences\\\": [ { \\\"type\\\": \\\"cjmiam\\\", \\\"detail\\\": { \\\"remoteAssets\\\": [], \\\"html\\\": \\\"<html><head><\\/head><body bgcolor=\\\\\\\"black\\\\\\\"><br \\/><br \\/><br \\/><br \\/><br \\/><br \\/><h1 align=\\\\\\\"center\\\\\\\" style=\\\\\\\"color: white;\\\\\\\">IN-APP MESSAGING POWERED BY <br \\/>OFFER DECISIONING<\\/h1><h1 align=\\\\\\\"center\\\\\\\"><a style=\\\\\\\"color: white;\\\\\\\" href=\\\\\\\"adbinapp:\\/\\/cancel\\\\\\\" >dismiss me<\\/a><\\/h1><\\/body><\\/html>\\\", \\\"template\\\": \\\"fullscreen\\\" } } ] } ] }\"\n" +
-                "          },\n" +
-                "          \"etag\" : \"1\",\n" +
-                "          \"schema\" : \"https:\\/\\/ns.adobe.com\\/experience\\/offer-management\\/content-component-json\"\n" +
-                "        }\n" +
-                "      ],\n" +
-                "      \"id\" : \"cb25ecb0-d085-44ac-b73d-797a3265d37c\"\n" +
-                "    }\n" +
-                "  ]");
-        eventData.put("payload", payload);
-        eventData.put("requestId", "D158979E-0506-4968-8031-17A6A8A87DA8");
-
-        // Mocks
-        Event mockEvent = mock(Event.class);
-
-        // when mock event getType called return EDGE
-        when(mockEvent.getType()).thenReturn(MessagingConstants.EventType.EDGE);
-
-        // when mock event getSource called return PERSONALIZATION_DECISIONS
-        when(mockEvent.getSource()).thenReturn(MessagingConstants.EventSource.PERSONALIZATION_DECISIONS);
-
-        // when get eventData called return event data containing an invalid offers iam payload
-        when(mockEvent.getEventData()).thenReturn(eventData);
-
-        // test
-        messagingInternal.queueEvent(mockEvent);
-        messagingInternal.processEvents();
-
-        // verify no rules loaded for messaging extension
-        ConcurrentHashMap loadedRules = mockCore.eventHub.getModuleRuleAssociation();
-        assertEquals(0, loadedRules.size());
-    }
-
-    @Test
-    public void test_handleEdgeResponseEvent_OffersIAMPayloadMissingMessageType() throws Exception {
-        // setup
-        // private mocks
-        Whitebox.setInternalState(messagingInternal, "messagingState", messagingState);
-        // trigger event
-        HashMap<String, Object> eventData = new HashMap<>();
-        eventData.put("type", "personalization:decisions");
-        eventData.put("requestEventId", "2E964037-E319-4D14-98B8-0682374E547B");
-        JSONObject payload = new JSONObject("    {\n" +
-                "      \"activity\" : {\n" +
-                "        \"id\" : \"mock_activity\",\n" +
-                "        \"etag\" : \"2\"\n" +
-                "      },\n" +
-                "      \"scope\" : \"eyJhY3Rpdml0eUlkIjoieGNvcmU6b2ZmZXItYWN0aXZpdHk6MTMyM2RiZTk0ZjJlZWY5MyIsInBsYWNlbWVudElkIjoieGNvcmU6b2ZmZXItcGxhY2VtZW50OjEzMjNkOWViNDNhYWNhZGEiLCJpdGVtQ291bnQiOjMwfQ==\",\n" +
-                "      \"placement\" : {\n" +
-                "        \"id\" : \"mock_placement\",\n" +
-                "        \"etag\" : \"1\"\n" +
-                "      },\n" +
-                "      \"items\" : [\n" +
-                "        {\n" +
-                "          \"id\" : \"xcore:fallback-offer:1323dbbc1c6eef91\",\n" +
-                "          \"data\" : {\n" +
-                "            \"id\" : \"xcore:fallback-offer:1323dbbc1c6eef91\",\n" +
-                "            \"format\" : \"application\\/json\",\n" +
-                "            \"content\" : \"{ \\\"version\\\": 1, \\\"rules\\\": [ { \\\"condition\\\": { \\\"type\\\": \\\"group\\\", \\\"definition\\\": { \\\"logic\\\": \\\"and\\\", \\\"conditions\\\": [ { \\\"definition\\\": { \\\"key\\\": \\\"contextdata.testShowMessage\\\", \\\"matcher\\\": \\\"eq\\\", \\\"values\\\": [ \\\"true\\\" ] }, \\\"type\\\": \\\"matcher\\\" } ] } }, \\\"consequences\\\": [ { \\\"id\\\": \\\"341800180\\\", \\\"detail\\\": { \\\"remoteAssets\\\": [], \\\"html\\\": \\\"<html><head><\\/head><body bgcolor=\\\\\\\"black\\\\\\\"><br \\/><br \\/><br \\/><br \\/><br \\/><br \\/><h1 align=\\\\\\\"center\\\\\\\" style=\\\\\\\"color: white;\\\\\\\">IN-APP MESSAGING POWERED BY <br \\/>OFFER DECISIONING<\\/h1><h1 align=\\\\\\\"center\\\\\\\"><a style=\\\\\\\"color: white;\\\\\\\" href=\\\\\\\"adbinapp:\\/\\/cancel\\\\\\\" >dismiss me<\\/a><\\/h1><\\/body><\\/html>\\\", \\\"template\\\": \\\"fullscreen\\\" } } ] } ] }\"\n" +
-                "          },\n" +
-                "          \"etag\" : \"1\",\n" +
-                "          \"schema\" : \"https:\\/\\/ns.adobe.com\\/experience\\/offer-management\\/content-component-json\"\n" +
-                "        }\n" +
-                "      ],\n" +
-                "      \"id\" : \"cb25ecb0-d085-44ac-b73d-797a3265d37c\"\n" +
-                "    }\n" +
-                "  ]");
-        eventData.put("payload", payload);
-        eventData.put("requestId", "D158979E-0506-4968-8031-17A6A8A87DA8");
-
-        // Mocks
-        Event mockEvent = mock(Event.class);
-
-        // when mock event getType called return EDGE
-        when(mockEvent.getType()).thenReturn(MessagingConstants.EventType.EDGE);
-
-        // when mock event getSource called return PERSONALIZATION_DECISIONS
-        when(mockEvent.getSource()).thenReturn(MessagingConstants.EventSource.PERSONALIZATION_DECISIONS);
-
-        // when get eventData called return event data containing an invalid offers iam payload
-        when(mockEvent.getEventData()).thenReturn(eventData);
-
-        // test
-        messagingInternal.queueEvent(mockEvent);
-        messagingInternal.processEvents();
-
-        // verify no rule loaded for messaging extension
-        ConcurrentHashMap loadedRules = mockCore.eventHub.getModuleRuleAssociation();
-        assertEquals(0, loadedRules.size());
-    }
-
-    @Test
-    public void test_handleEdgeResponseEvent_OffersIAMPayloadMissingMessageDetail() throws Exception {
-        // setup
-        // private mocks
-        Whitebox.setInternalState(messagingInternal, "messagingState", messagingState);
-        // trigger event
-        HashMap<String, Object> eventData = new HashMap<>();
-        eventData.put("type", "personalization:decisions");
-        eventData.put("requestEventId", "2E964037-E319-4D14-98B8-0682374E547B");
-        JSONObject payload = new JSONObject("    {\n" +
-                "      \"activity\" : {\n" +
-                "        \"id\" : \"mock_activity\",\n" +
-                "        \"etag\" : \"2\"\n" +
-                "      },\n" +
-                "      \"scope\" : \"eyJhY3Rpdml0eUlkIjoieGNvcmU6b2ZmZXItYWN0aXZpdHk6MTMyM2RiZTk0ZjJlZWY5MyIsInBsYWNlbWVudElkIjoieGNvcmU6b2ZmZXItcGxhY2VtZW50OjEzMjNkOWViNDNhYWNhZGEiLCJpdGVtQ291bnQiOjMwfQ==\",\n" +
-                "      \"placement\" : {\n" +
-                "        \"id\" : \"mock_placement\",\n" +
-                "        \"etag\" : \"1\"\n" +
-                "      },\n" +
-                "      \"items\" : [\n" +
-                "        {\n" +
-                "          \"id\" : \"xcore:fallback-offer:1323dbbc1c6eef91\",\n" +
-                "          \"data\" : {\n" +
-                "            \"id\" : \"xcore:fallback-offer:1323dbbc1c6eef91\",\n" +
-                "            \"format\" : \"application\\/json\",\n" +
-                "            \"content\" : \"{ \\\"version\\\": 1, \\\"rules\\\": [ { \\\"condition\\\": { \\\"type\\\": \\\"group\\\", \\\"definition\\\": { \\\"logic\\\": \\\"and\\\", \\\"conditions\\\": [ { \\\"definition\\\": { \\\"key\\\": \\\"contextdata.testShowMessage\\\", \\\"matcher\\\": \\\"eq\\\", \\\"values\\\": [ \\\"true\\\" ] }, \\\"type\\\": \\\"matcher\\\" } ] } }, \\\"consequences\\\": [ { \\\"id\\\": \\\"341800180\\\", \\\"type\\\": \\\"cjmiam\\\"} ] } ] }\"\n" +
-                "          },\n" +
-                "          \"etag\" : \"1\",\n" +
-                "          \"schema\" : \"https:\\/\\/ns.adobe.com\\/experience\\/offer-management\\/content-component-json\"\n" +
-                "        }\n" +
-                "      ],\n" +
-                "      \"id\" : \"cb25ecb0-d085-44ac-b73d-797a3265d37c\"\n" +
-                "    }\n" +
-                "  ]");
-        eventData.put("payload", payload);
-        eventData.put("requestId", "D158979E-0506-4968-8031-17A6A8A87DA8");
-
-        // Mocks
-        Event mockEvent = mock(Event.class);
-
-        // when mock event getType called return EDGE
-        when(mockEvent.getType()).thenReturn(MessagingConstants.EventType.EDGE);
-
-        // when mock event getSource called return PERSONALIZATION_DECISIONS
-        when(mockEvent.getSource()).thenReturn(MessagingConstants.EventSource.PERSONALIZATION_DECISIONS);
-
-        // when get eventData called return event data containing an invalid offers iam payload
-        when(mockEvent.getEventData()).thenReturn(eventData);
-
-        // test
-        messagingInternal.queueEvent(mockEvent);
-        messagingInternal.processEvents();
-
-        // verify no rule loaded for messaging extension
-        ConcurrentHashMap loadedRules = mockCore.eventHub.getModuleRuleAssociation();
-        assertEquals(0, loadedRules.size());
-    }
-
-    @Test
-    public void test_handleEdgeResponseEvent_OffersIAMPayloadIsEmpty() throws Exception {
-        // setup
-        // private mocks
-        Whitebox.setInternalState(messagingInternal, "messagingState", messagingState);
-        // trigger event
-        HashMap<String, Object> eventData = new HashMap<>();
-        eventData.put("type", "personalization:decisions");
-        eventData.put("requestEventId", "2E964037-E319-4D14-98B8-0682374E547B");
-        JSONObject payload = new JSONObject("    {\n" +
-                "      \"activity\" : {\n" +
-                "        \"id\" : \"mock_activity\",\n" +
-                "        \"etag\" : \"2\"\n" +
-                "      },\n" +
-                "      \"scope\" : \"eyJhY3Rpdml0eUlkIjoieGNvcmU6b2ZmZXItYWN0aXZpdHk6MTMyM2RiZTk0ZjJlZWY5MyIsInBsYWNlbWVudElkIjoieGNvcmU6b2ZmZXItcGxhY2VtZW50OjEzMjNkOWViNDNhYWNhZGEiLCJpdGVtQ291bnQiOjMwfQ==\",\n" +
-                "      \"placement\" : {\n" +
-                "        \"id\" : \"mock_placement\",\n" +
-                "        \"etag\" : \"1\"\n" +
-                "      },\n" +
-                "      \"items\" : [\n" +
-                "        {\n" +
-                "          \"id\" : \"xcore:fallback-offer:1323dbbc1c6eef91\",\n" +
-                "          \"data\" : {\n" +
-                "            \"id\" : \"xcore:fallback-offer:1323dbbc1c6eef91\",\n" +
-                "            \"format\" : \"application\\/json\",\n" +
-                "            \"content\" : \"{ \\\"version\\\": 1, \\\"rules\\\": [ { \\\"condition\\\": { \\\"type\\\": \\\"group\\\", \\\"definition\\\": { \\\"logic\\\": \\\"and\\\", \\\"conditions\\\": [ { \\\"definition\\\": { \\\"key\\\": \\\"contextdata.testShowMessage\\\", \\\"matcher\\\": \\\"eq\\\", \\\"values\\\": [ \\\"true\\\" ] }, \\\"type\\\": \\\"matcher\\\" } ] } }, \\\"consequences\\\": [ { \\\"id\\\": \\\"341800180\\\", \\\"type\\\": \\\"cjmiam\\\"} ] } ] }\"\n" +
-                "          },\n" +
-                "          \"etag\" : \"1\",\n" +
-                "          \"schema\" : \"https:\\/\\/ns.adobe.com\\/experience\\/offer-management\\/content-component-json\"\n" +
-                "        }\n" +
-                "      ],\n" +
-                "      \"id\" : \"cb25ecb0-d085-44ac-b73d-797a3265d37c\"\n" +
-                "    }\n" +
-                "  ]");
-        eventData.put("payload", payload);
-        eventData.put("requestId", "D158979E-0506-4968-8031-17A6A8A87DA8");
-
-        // Mocks
-        Event mockEvent = mock(Event.class);
-
-        // when mock event getType called return EDGE
-        when(mockEvent.getType()).thenReturn(MessagingConstants.EventType.EDGE);
-
-        // when mock event getSource called return PERSONALIZATION_DECISIONS
-        when(mockEvent.getSource()).thenReturn(MessagingConstants.EventSource.PERSONALIZATION_DECISIONS);
-
-        // when get eventData called return event data containing an invalid offers iam payload
-        when(mockEvent.getEventData()).thenReturn(eventData);
-
-        // test
-        messagingInternal.queueEvent(mockEvent);
-        messagingInternal.processEvents();
-
-        // verify no rule loaded for messaging extension
-        ConcurrentHashMap loadedRules = mockCore.eventHub.getModuleRuleAssociation();
-        assertEquals(0, loadedRules.size());
-    }
-
-    @Test
-    public void test_handleEdgeResponseEvent_OffersIAMPayloadHasInvalidActivityId() throws Exception {
-        // setup
-        // private mocks
-        Whitebox.setInternalState(messagingInternal, "messagingState", messagingState);
-        // trigger event
-        HashMap<String, Object> eventData = new HashMap<>();
-        eventData.put("type", "personalization:decisions");
-        eventData.put("requestEventId", "2E964037-E319-4D14-98B8-0682374E547B");
-        JSONObject payload = new JSONObject("    {\n" +
-                "      \"activity\" : {\n" +
-                "        \"id\" : \"invalid\",\n" +
-                "        \"etag\" : \"2\"\n" +
-                "      },\n" +
-                "      \"scope\" : \"eyJhY3Rpdml0eUlkIjoieGNvcmU6b2ZmZXItYWN0aXZpdHk6MTMyM2RiZTk0ZjJlZWY5MyIsInBsYWNlbWVudElkIjoieGNvcmU6b2ZmZXItcGxhY2VtZW50OjEzMjNkOWViNDNhYWNhZGEiLCJpdGVtQ291bnQiOjMwfQ==\",\n" +
-                "      \"placement\" : {\n" +
-                "        \"id\" : \"mock_placement\",\n" +
-                "        \"etag\" : \"1\"\n" +
-                "      },\n" +
-                "      \"items\" : [\n" +
-                "        {\n" +
-                "          \"id\" : \"xcore:fallback-offer:1323dbbc1c6eef91\",\n" +
-                "          \"data\" : {\n" +
-                "            \"id\" : \"xcore:fallback-offer:1323dbbc1c6eef91\",\n" +
-                "            \"format\" : \"application\\/json\",\n" +
-                "            \"content\" : \"{ \\\"version\\\": 1, \\\"rules\\\": [ { \\\"condition\\\": { \\\"type\\\": \\\"group\\\", \\\"definition\\\": { \\\"logic\\\": \\\"and\\\", \\\"conditions\\\": [ { \\\"definition\\\": { \\\"key\\\": \\\"contextdata.testShowMessage\\\", \\\"matcher\\\": \\\"eq\\\", \\\"values\\\": [ \\\"true\\\" ] }, \\\"type\\\": \\\"matcher\\\" } ] } }, \\\"consequences\\\": [ { \\\"id\\\": \\\"341800180\\\", \\\"type\\\": \\\"cjmiam\\\"} ] } ] }\"\n" +
-                "          },\n" +
-                "          \"etag\" : \"1\",\n" +
-                "          \"schema\" : \"https:\\/\\/ns.adobe.com\\/experience\\/offer-management\\/content-component-json\"\n" +
-                "        }\n" +
-                "      ],\n" +
-                "      \"id\" : \"cb25ecb0-d085-44ac-b73d-797a3265d37c\"\n" +
-                "    }\n" +
-                "  ]");
-        eventData.put("payload", payload);
-        eventData.put("requestId", "D158979E-0506-4968-8031-17A6A8A87DA8");
-
-        // Mocks
-        Event mockEvent = mock(Event.class);
-
-        // when mock event getType called return EDGE
-        when(mockEvent.getType()).thenReturn(MessagingConstants.EventType.EDGE);
-
-        // when mock event getSource called return PERSONALIZATION_DECISIONS
-        when(mockEvent.getSource()).thenReturn(MessagingConstants.EventSource.PERSONALIZATION_DECISIONS);
-
-        // when get eventData called return event data containing an invalid offers iam payload
-        when(mockEvent.getEventData()).thenReturn(eventData);
-
-        // test
-        messagingInternal.queueEvent(mockEvent);
-        messagingInternal.processEvents();
-
-        // verify no rule loaded for messaging extension
-        ConcurrentHashMap loadedRules = mockCore.eventHub.getModuleRuleAssociation();
-        assertEquals(0, loadedRules.size());
-    }
-
-    @Test
-    public void test_handleEdgeResponseEvent_OffersIAMPayloadHasInvalidPlacementId() throws Exception {
-        // setup
-        // private mocks
-        Whitebox.setInternalState(messagingInternal, "messagingState", messagingState);
-        // trigger event
-        HashMap<String, Object> eventData = new HashMap<>();
-        eventData.put("type", "personalization:decisions");
-        eventData.put("requestEventId", "2E964037-E319-4D14-98B8-0682374E547B");
-        JSONObject payload = new JSONObject("    {\n" +
-                "      \"activity\" : {\n" +
-                "        \"id\" : \"mock_activity\",\n" +
-                "        \"etag\" : \"2\"\n" +
-                "      },\n" +
-                "      \"scope\" : \"eyJhY3Rpdml0eUlkIjoieGNvcmU6b2ZmZXItYWN0aXZpdHk6MTMyM2RiZTk0ZjJlZWY5MyIsInBsYWNlbWVudElkIjoieGNvcmU6b2ZmZXItcGxhY2VtZW50OjEzMjNkOWViNDNhYWNhZGEiLCJpdGVtQ291bnQiOjMwfQ==\",\n" +
-                "      \"placement\" : {\n" +
-                "        \"id\" : \"invalid\",\n" +
-                "        \"etag\" : \"1\"\n" +
-                "      },\n" +
-                "      \"items\" : [\n" +
-                "        {\n" +
-                "          \"id\" : \"xcore:fallback-offer:1323dbbc1c6eef91\",\n" +
-                "          \"data\" : {\n" +
-                "            \"id\" : \"xcore:fallback-offer:1323dbbc1c6eef91\",\n" +
-                "            \"format\" : \"application\\/json\",\n" +
-                "            \"content\" : \"{ \\\"version\\\": 1, \\\"rules\\\": [ { \\\"condition\\\": { \\\"type\\\": \\\"group\\\", \\\"definition\\\": { \\\"logic\\\": \\\"and\\\", \\\"conditions\\\": [ { \\\"definition\\\": { \\\"key\\\": \\\"contextdata.testShowMessage\\\", \\\"matcher\\\": \\\"eq\\\", \\\"values\\\": [ \\\"true\\\" ] }, \\\"type\\\": \\\"matcher\\\" } ] } }, \\\"consequences\\\": [ { \\\"id\\\": \\\"341800180\\\", \\\"type\\\": \\\"cjmiam\\\"} ] } ] }\"\n" +
-                "          },\n" +
-                "          \"etag\" : \"1\",\n" +
-                "          \"schema\" : \"https:\\/\\/ns.adobe.com\\/experience\\/offer-management\\/content-component-json\"\n" +
-                "        }\n" +
-                "      ],\n" +
-                "      \"id\" : \"cb25ecb0-d085-44ac-b73d-797a3265d37c\"\n" +
-                "    }\n" +
-                "  ]");
-        eventData.put("payload", payload);
-        eventData.put("requestId", "D158979E-0506-4968-8031-17A6A8A87DA8");
-
-        // Mocks
-        Event mockEvent = mock(Event.class);
-
-        // when mock event getType called return EDGE
-        when(mockEvent.getType()).thenReturn(MessagingConstants.EventType.EDGE);
-
-        // when mock event getSource called return PERSONALIZATION_DECISIONS
-        when(mockEvent.getSource()).thenReturn(MessagingConstants.EventSource.PERSONALIZATION_DECISIONS);
-
-        // when get eventData called return event data containing an invalid offers iam payload
-        when(mockEvent.getEventData()).thenReturn(eventData);
-
-        // test
-        messagingInternal.queueEvent(mockEvent);
-        messagingInternal.processEvents();
-
-        // verify no rule loaded for messaging extension
-        ConcurrentHashMap loadedRules = mockCore.eventHub.getModuleRuleAssociation();
-        assertEquals(0, loadedRules.size());
-    }
-    // ========================================================================================
-    // handling rules response events
-    // ========================================================================================
-    @Test
-    public void test_handleRulesResponseEvent_ValidConsequencePresent() {
-        // setup
-        MobileCore.setApplication(App.getApplication());
-        // private mocks
-        Whitebox.setInternalState(messagingInternal, "messagingState", messagingState);
-        Mockito.when(mockUIService.createFullscreenMessage(any(String.class), any(UIService.FullscreenMessageDelegate.class), any(boolean.class), any(Object.class))).thenReturn(mockFullscreenMessage);
-        // trigger event
-        Map<String, Object> consequence = new HashMap<>();
-        Map<String, Object> details = new HashMap<>();
-        details.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL_KEY_TEMPLATE, "fullscreen");
-        details.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL_KEY_REMOTE_ASSETS, new ArrayList<String>());
-        details.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL_KEY_HTML, "<html><head></head><body bgcolor=\"black\"><br /><br /><br /><br /><br /><br /><h1 align=\"center\" style=\"color: white;\">IN-APP MESSAGING POWERED BY <br />OFFER DECISIONING</h1><h1 align=\"center\"><a style=\"color: white;\" href=\"adbinapp://cancel\" >dismiss me</a></h1></body></html>");
-        consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_ID, "123456789");
-        consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL, details);
-        consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_TYPE, MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_CJM_VALUE);
-        Map<String, Object> eventData = new HashMap<>();
-        eventData.put(MessagingConstants.EventDataKeys.RulesEngine.CONSEQUENCE_TRIGGERED, consequence);
-
-        // Mocks
-        Event mockEvent = mock(Event.class);
-
-        // when mock event getType called return RULES_ENGINE
-        when(mockEvent.getType()).thenReturn(EventType.RULES_ENGINE.getName());
-
-        // when mock event getSource called return RESPONSE_CONTENT
-        when(mockEvent.getSource()).thenReturn(EventSource.RESPONSE_CONTENT.getName());
-
-        // when get eventData called return event data containing a valid rules response content event with a triggered consequence
-        when(mockEvent.getEventData()).thenReturn(eventData);
-
-        // test
-        messagingInternal.queueEvent(mockEvent);
-        messagingInternal.processEvents();
-
-        // verify MessagingFullscreenMessage.show() is called
-        verify(mockFullscreenMessage, times(1)).show();
-    }
-
-    @Test
-    public void test_handleRulesResponseEvent_UnsupportedConsequenceType() {
-        // setup
-        MobileCore.setApplication(App.getApplication());
-        // private mocks
-        Whitebox.setInternalState(messagingInternal, "messagingState", messagingState);
-                Mockito.when(mockUIService.createFullscreenMessage(any(String.class), any(UIService.FullscreenMessageDelegate.class), any(boolean.class), any(Extension.class))).thenReturn(mockFullscreenMessage);
-
-        // trigger event
-        Map<String, Object> consequence = new HashMap<>();
-        Map<String, Object> details = new HashMap<>();
-        details.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL_KEY_TEMPLATE, "fullscreen");
-        details.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL_KEY_REMOTE_ASSETS, new ArrayList<String>());
-        details.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL_KEY_HTML, "<html><head></head><body bgcolor=\"black\"><br /><br /><br /><br /><br /><br /><h1 align=\"center\" style=\"color: white;\">IN-APP MESSAGING POWERED BY <br />OFFER DECISIONING</h1><h1 align=\"center\"><a style=\"color: white;\" href=\"adbinapp://cancel\" >dismiss me</a></h1></body></html>");
-        consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_ID, "123456789");
-        consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL, details);
-        consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_TYPE, "unknownIamType");
-        Map<String, Object> eventData = new HashMap<>();
-        eventData.put(MessagingConstants.EventDataKeys.RulesEngine.CONSEQUENCE_TRIGGERED, consequence);
-
-        // Mocks
-        Event mockEvent = mock(Event.class);
-
-        // when mock event getType called return RULES_ENGINE
-        when(mockEvent.getType()).thenReturn(EventType.RULES_ENGINE.getName());
-
-        // when mock event getSource called return RESPONSE_CONTENT
-        when(mockEvent.getSource()).thenReturn(EventSource.RESPONSE_CONTENT.getName());
-
-        // when get eventData called return event data containing a rules response content event with an invalid triggered consequence
-        when(mockEvent.getEventData()).thenReturn(eventData);
-
-        // test
-        messagingInternal.queueEvent(mockEvent);
-        messagingInternal.processEvents();
-
-        // verify MessagingFullscreenMessage.show() is not called
-        verify(mockFullscreenMessage, times(0)).show();
-    }
-
-    @Test
-    public void test_handleRulesResponseEvent_MissingConsequenceType() {
-        // setup
-        MobileCore.setApplication(App.getApplication());
-        // private mocks
-        Whitebox.setInternalState(messagingInternal, "messagingState", messagingState);
-                Mockito.when(mockUIService.createFullscreenMessage(any(String.class), any(UIService.FullscreenMessageDelegate.class), any(boolean.class), any(Extension.class))).thenReturn(mockFullscreenMessage);
-
-        // trigger event
-        Map<String, Object> consequence = new HashMap<>();
-        Map<String, Object> details = new HashMap<>();
-        details.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL_KEY_TEMPLATE, "fullscreen");
-        details.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL_KEY_REMOTE_ASSETS, new ArrayList<String>());
-        details.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL_KEY_HTML, "<html><head></head><body bgcolor=\"black\"><br /><br /><br /><br /><br /><br /><h1 align=\"center\" style=\"color: white;\">IN-APP MESSAGING POWERED BY <br />OFFER DECISIONING</h1><h1 align=\"center\"><a style=\"color: white;\" href=\"adbinapp://cancel\" >dismiss me</a></h1></body></html>");
-        consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_ID, "123456789");
-        consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL, details);
-        Map<String, Object> eventData = new HashMap<>();
-        eventData.put(MessagingConstants.EventDataKeys.RulesEngine.CONSEQUENCE_TRIGGERED, consequence);
-
-        // Mocks
-        Event mockEvent = mock(Event.class);
-
-        // when mock event getType called return RULES_ENGINE
-        when(mockEvent.getType()).thenReturn(EventType.RULES_ENGINE.getName());
-
-        // when mock event getSource called return RESPONSE_CONTENT
-        when(mockEvent.getSource()).thenReturn(EventSource.RESPONSE_CONTENT.getName());
-
-        // when get eventData called return event data containing a rules response content event with an invalid triggered consequence
-        when(mockEvent.getEventData()).thenReturn(eventData);
-
-        // test
-        messagingInternal.queueEvent(mockEvent);
-        messagingInternal.processEvents();
-
-        // verify MessagingFullscreenMessage.show() is not called
-        verify(mockFullscreenMessage, times(0)).show();
-    }
-
-    public void test_handleRulesResponseEvent_NullConsequences() {
-        // setup
-        MobileCore.setApplication(App.getApplication());
-        // private mocks
-        Whitebox.setInternalState(messagingInternal, "messagingState", messagingState);
-                Mockito.when(mockUIService.createFullscreenMessage(any(String.class), any(UIService.FullscreenMessageDelegate.class), any(boolean.class), any(Extension.class))).thenReturn(mockFullscreenMessage);
-
-        // trigger event
-        Map<String, Object> eventData = new HashMap<>();
-        eventData.put(MessagingConstants.EventDataKeys.RulesEngine.CONSEQUENCE_TRIGGERED, null);
-
-        // Mocks
-        Event mockEvent = mock(Event.class);
-
-        // when mock event getType called return RULES_ENGINE
-        when(mockEvent.getType()).thenReturn(EventType.RULES_ENGINE.getName());
-
-        // when mock event getSource called return RESPONSE_CONTENT
-        when(mockEvent.getSource()).thenReturn(EventSource.RESPONSE_CONTENT.getName());
-
-        // when get eventData called return event data containing a rules response content event with an invalid triggered consequence
-        when(mockEvent.getEventData()).thenReturn(eventData);
-
-        // test
-        messagingInternal.queueEvent(mockEvent);
-        messagingInternal.processEvents();
-
-        // verify MessagingFullscreenMessage.show() is not called
-        verify(mockFullscreenMessage, times(0)).show();
-    }
-
-    @Test
-    public void test_handleRulesResponseEvent_InvalidConsequenceDetails_EmptyConsequences() {
-        // setup
-        MobileCore.setApplication(App.getApplication());
-        // private mocks
-        Whitebox.setInternalState(messagingInternal, "messagingState", messagingState);
-                Mockito.when(mockUIService.createFullscreenMessage(any(String.class), any(UIService.FullscreenMessageDelegate.class), any(boolean.class), any(Extension.class))).thenReturn(mockFullscreenMessage);
-
-        // trigger event
-        Map<String, Object> consequence = new HashMap<>();
-        Map<String, Object> eventData = new HashMap<>();
-        eventData.put(MessagingConstants.EventDataKeys.RulesEngine.CONSEQUENCE_TRIGGERED, consequence);
-
-        // Mocks
-        Event mockEvent = mock(Event.class);
-
-        // when mock event getType called return RULES_ENGINE
-        when(mockEvent.getType()).thenReturn(EventType.RULES_ENGINE.getName());
-
-        // when mock event getSource called return RESPONSE_CONTENT
-        when(mockEvent.getSource()).thenReturn(EventSource.RESPONSE_CONTENT.getName());
-
-        // when get eventData called return event data containing a rules response content event with an invalid triggered consequence
-        when(mockEvent.getEventData()).thenReturn(eventData);
-
-        // test
-        messagingInternal.queueEvent(mockEvent);
-        messagingInternal.processEvents();
-
-        // verify MessagingFullscreenMessage.show() is not called
-        verify(mockFullscreenMessage, times(0)).show();
-    }
-
-    @Test
-    public void test_handleRulesResponseEvent_EmptyDetails() {
-        // setup
-        MobileCore.setApplication(App.getApplication());
-        // private mocks
-        Whitebox.setInternalState(messagingInternal, "messagingState", messagingState);
-                Mockito.when(mockUIService.createFullscreenMessage(any(String.class), any(UIService.FullscreenMessageDelegate.class), any(boolean.class), any(Extension.class))).thenReturn(mockFullscreenMessage);
-
-        // trigger event
-        Map<String, Object> eventData = new HashMap<>();
-        Map<String, Object> consequence = new HashMap<>();
-        Map<String, Object> details = new HashMap<>();
-        consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_ID, "123456789");
-        consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL, details);
-        consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_TYPE, MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_CJM_VALUE);
-        eventData.put(MessagingConstants.EventDataKeys.RulesEngine.CONSEQUENCE_TRIGGERED, consequence);
-
-        // Mocks
-        Event mockEvent = mock(Event.class);
-
-        // when mock event getType called return RULES_ENGINE
-        when(mockEvent.getType()).thenReturn(EventType.RULES_ENGINE.getName());
-
-        // when mock event getSource called return RESPONSE_CONTENT
-        when(mockEvent.getSource()).thenReturn(EventSource.RESPONSE_CONTENT.getName());
-
-        // when get eventData called return event data containing a rules response content event with an invalid triggered consequence
-        when(mockEvent.getEventData()).thenReturn(eventData);
-
-        // test
-        messagingInternal.queueEvent(mockEvent);
-        messagingInternal.processEvents();
-
-        // verify MessagingFullscreenMessage.show() is not called
-        verify(mockFullscreenMessage, times(0)).show();
-    }
-
-    @Test
-    public void test_handleRulesResponseEvent_NullDetails() {
-        // setup
-        MobileCore.setApplication(App.getApplication());
-        // private mocks
-        Whitebox.setInternalState(messagingInternal, "messagingState", messagingState);
-                Mockito.when(mockUIService.createFullscreenMessage(any(String.class), any(UIService.FullscreenMessageDelegate.class), any(boolean.class), any(Extension.class))).thenReturn(mockFullscreenMessage);
-
-        // trigger event
-        Map<String, Object> eventData = new HashMap<>();
-        Map<String, Object> consequence = new HashMap<>();
-        consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_ID, "123456789");
-        consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL, null);
-        consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_TYPE, MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_CJM_VALUE);
-        eventData.put(MessagingConstants.EventDataKeys.RulesEngine.CONSEQUENCE_TRIGGERED, consequence);
-
-        // Mocks
-        Event mockEvent = mock(Event.class);
-
-        // when mock event getType called return RULES_ENGINE
-        when(mockEvent.getType()).thenReturn(EventType.RULES_ENGINE.getName());
-
-        // when mock event getSource called return RESPONSE_CONTENT
-        when(mockEvent.getSource()).thenReturn(EventSource.RESPONSE_CONTENT.getName());
-
-        // when get eventData called return event data containing a rules response content event with an invalid triggered consequence
-        when(mockEvent.getEventData()).thenReturn(eventData);
-
-        // test
-        messagingInternal.queueEvent(mockEvent);
-        messagingInternal.processEvents();
-
-        // verify MessagingFullscreenMessage.show() is not called
-        verify(mockFullscreenMessage, times(0)).show();
-    }
-
-    @Test
-    public void test_handleRulesResponseEvent_InvalidConsequenceDetails_TemplateUnsupported() {
-        // setup
-        MobileCore.setApplication(App.getApplication());
-        // private mocks
-        Whitebox.setInternalState(messagingInternal, "messagingState", messagingState);
-                Mockito.when(mockUIService.createFullscreenMessage(any(String.class), any(UIService.FullscreenMessageDelegate.class), any(boolean.class), any(Extension.class))).thenReturn(mockFullscreenMessage);
-
-        // trigger event
-        Map<String, Object> consequence = new HashMap<>();
-        Map<String, Object> details = new HashMap<>();
-        details.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL_KEY_TEMPLATE, "alert");
-        details.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL_KEY_REMOTE_ASSETS, new ArrayList<String>());
-        details.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL_KEY_HTML, "<html><head></head><body bgcolor=\"black\"><br /><br /><br /><br /><br /><br /><h1 align=\"center\" style=\"color: white;\">IN-APP MESSAGING POWERED BY <br />OFFER DECISIONING</h1><h1 align=\"center\"><a style=\"color: white;\" href=\"adbinapp://cancel\" >dismiss me</a></h1></body></html>");
-        consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_ID, "123456789");
-        consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL, details);
-        consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_TYPE, MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_CJM_VALUE);
-        Map<String, Object> eventData = new HashMap<>();
-        eventData.put(MessagingConstants.EventDataKeys.RulesEngine.CONSEQUENCE_TRIGGERED, consequence);
-
-        // Mocks
-        Event mockEvent = mock(Event.class);
-
-        // when mock event getType called return RULES_ENGINE
-        when(mockEvent.getType()).thenReturn(EventType.RULES_ENGINE.getName());
-
-        // when mock event getSource called return RESPONSE_CONTENT
-        when(mockEvent.getSource()).thenReturn(EventSource.RESPONSE_CONTENT.getName());
-
-        // when get eventData called return event data containing a rules response content event with an invalid triggered consequence
-        when(mockEvent.getEventData()).thenReturn(eventData);
-
-        // test
-        messagingInternal.queueEvent(mockEvent);
-        messagingInternal.processEvents();
-
-        // verify MessagingFullscreenMessage.show() is not called
-        verify(mockFullscreenMessage, times(0)).show();
-    }
-
-    @Test
-    public void test_handleRulesResponseEvent_InvalidConsequenceDetails_TemplateMissing() {
-        // setup
-        MobileCore.setApplication(App.getApplication());
-        // private mocks
-        Whitebox.setInternalState(messagingInternal, "messagingState", messagingState);
-                Mockito.when(mockUIService.createFullscreenMessage(any(String.class), any(UIService.FullscreenMessageDelegate.class), any(boolean.class), any(Extension.class))).thenReturn(mockFullscreenMessage);
-
-        // trigger event
-        Map<String, Object> consequence = new HashMap<>();
-        Map<String, Object> details = new HashMap<>();
-        details.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL_KEY_REMOTE_ASSETS, new ArrayList<String>());
-        details.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL_KEY_HTML, "<html><head></head><body bgcolor=\"black\"><br /><br /><br /><br /><br /><br /><h1 align=\"center\" style=\"color: white;\">IN-APP MESSAGING POWERED BY <br />OFFER DECISIONING</h1><h1 align=\"center\"><a style=\"color: white;\" href=\"adbinapp://cancel\" >dismiss me</a></h1></body></html>");
-        consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_ID, "123456789");
-        consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL, details);
-        consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_TYPE, MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_CJM_VALUE);
-        Map<String, Object> eventData = new HashMap<>();
-        eventData.put(MessagingConstants.EventDataKeys.RulesEngine.CONSEQUENCE_TRIGGERED, consequence);
-
-        // Mocks
-        Event mockEvent = mock(Event.class);
-
-        // when mock event getType called return RULES_ENGINE
-        when(mockEvent.getType()).thenReturn(EventType.RULES_ENGINE.getName());
-
-        // when mock event getSource called return RESPONSE_CONTENT
-        when(mockEvent.getSource()).thenReturn(EventSource.RESPONSE_CONTENT.getName());
-
-        // when get eventData called return event data containing a rules response content event with an invalid triggered consequence
-        when(mockEvent.getEventData()).thenReturn(eventData);
-
-        // test
-        messagingInternal.queueEvent(mockEvent);
-        messagingInternal.processEvents();
-
-        // verify MessagingFullscreenMessage.show() is not called
-        verify(mockFullscreenMessage, times(0)).show();
-    }
-
-    @Test
-    public void test_handleRulesResponseEvent_InvalidConsequenceDetails_EmptyHTML() {
-        // setup
-        MobileCore.setApplication(App.getApplication());
-        // private mocks
-        Whitebox.setInternalState(messagingInternal, "messagingState", messagingState);
-                Mockito.when(mockUIService.createFullscreenMessage(any(String.class), any(UIService.FullscreenMessageDelegate.class), any(boolean.class), any(Extension.class))).thenReturn(mockFullscreenMessage);
-
-        // trigger event
-        Map<String, Object> consequence = new HashMap<>();
-        Map<String, Object> details = new HashMap<>();
-        details.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL_KEY_TEMPLATE, "fullscreen");
-        details.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL_KEY_REMOTE_ASSETS, new ArrayList<String>());
-        details.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL_KEY_HTML, "");
-                consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_ID, "123456789");
-        consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL, details);
-        consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_TYPE, MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_CJM_VALUE);
-        Map<String, Object> eventData = new HashMap<>();
-        eventData.put(MessagingConstants.EventDataKeys.RulesEngine.CONSEQUENCE_TRIGGERED, consequence);
-
-        // Mocks
-        Event mockEvent = mock(Event.class);
-
-        // when mock event getType called return RULES_ENGINE
-        when(mockEvent.getType()).thenReturn(EventType.RULES_ENGINE.getName());
-
-        // when mock event getSource called return RESPONSE_CONTENT
-        when(mockEvent.getSource()).thenReturn(EventSource.RESPONSE_CONTENT.getName());
-
-        // when get eventData called return event data containing a rules response content event with an invalid triggered consequence
-        when(mockEvent.getEventData()).thenReturn(eventData);
-
-        // test
-        messagingInternal.queueEvent(mockEvent);
-        messagingInternal.processEvents();
-
-        // verify MessagingFullscreenMessage.show() is not called
-        verify(mockFullscreenMessage, times(0)).show();
-    }
-
-    @Test
-    public void test_handleRulesResponseEvent_InvalidConsequenceDetails_MissingHTML() {
-        // setup
-        MobileCore.setApplication(App.getApplication());
-        // private mocks
-        Whitebox.setInternalState(messagingInternal, "messagingState", messagingState);
-                Mockito.when(mockUIService.createFullscreenMessage(any(String.class), any(UIService.FullscreenMessageDelegate.class), any(boolean.class), any(Extension.class))).thenReturn(mockFullscreenMessage);
-
-        // trigger event
-        Map<String, Object> consequence = new HashMap<>();
-        Map<String, Object> details = new HashMap<>();
-        details.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL_KEY_TEMPLATE, "fullscreen");
-        details.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL_KEY_REMOTE_ASSETS, new ArrayList<String>());
-        consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_ID, "123456789");
-        consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL, details);
-        consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_TYPE, MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_CJM_VALUE);
-        Map<String, Object> eventData = new HashMap<>();
-        eventData.put(MessagingConstants.EventDataKeys.RulesEngine.CONSEQUENCE_TRIGGERED, consequence);
-
-        // Mocks
-        Event mockEvent = mock(Event.class);
-
-        // when mock event getType called return RULES_ENGINE
-        when(mockEvent.getType()).thenReturn(EventType.RULES_ENGINE.getName());
-
-        // when mock event getSource called return RESPONSE_CONTENT
-        when(mockEvent.getSource()).thenReturn(EventSource.RESPONSE_CONTENT.getName());
-
-        // when get eventData called return event data containing a rules response content event with an invalid triggered consequence
-        when(mockEvent.getEventData()).thenReturn(eventData);
-
-        // test
-        messagingInternal.queueEvent(mockEvent);
-        messagingInternal.processEvents();
-
-        // verify MessagingFullscreenMessage.show() is not called
-        verify(mockFullscreenMessage, times(0)).show();
-    }
-
-    @Test
-    public void test_handleRulesResponseEvent_InvalidConsequenceDetails_MissingMessageId() {
-        // setup
-        MobileCore.setApplication(App.getApplication());
-        // private mocks
-        Whitebox.setInternalState(messagingInternal, "messagingState", messagingState);
-                Mockito.when(mockUIService.createFullscreenMessage(any(String.class), any(UIService.FullscreenMessageDelegate.class), any(boolean.class), any(Extension.class))).thenReturn(mockFullscreenMessage);
-
-        // trigger event
-        Map<String, Object> consequence = new HashMap<>();
-        Map<String, Object> details = new HashMap<>();
-        details.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL_KEY_TEMPLATE, "fullscreen");
-        details.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL_KEY_REMOTE_ASSETS, new ArrayList<String>());
-        consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_DETAIL, details);
-        consequence.put(MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_TYPE, MessagingConstants.EventDataKeys.RulesEngine.MESSAGE_CONSEQUENCE_CJM_VALUE);
-        Map<String, Object> eventData = new HashMap<>();
-        eventData.put(MessagingConstants.EventDataKeys.RulesEngine.CONSEQUENCE_TRIGGERED, consequence);
-
-        // Mocks
-        Event mockEvent = mock(Event.class);
-
-        // when mock event getType called return RULES_ENGINE
-        when(mockEvent.getType()).thenReturn(EventType.RULES_ENGINE.getName());
-
-        // when mock event getSource called return RESPONSE_CONTENT
-        when(mockEvent.getSource()).thenReturn(EventSource.RESPONSE_CONTENT.getName());
-
-        // when get eventData called return event data containing a rules response content event with an invalid triggered consequence
-        when(mockEvent.getEventData()).thenReturn(eventData);
-
-        // test
-        messagingInternal.queueEvent(mockEvent);
-        messagingInternal.processEvents();
-
-        // verify MessagingFullscreenMessage.show() is not called
-        verify(mockFullscreenMessage, times(0)).show();
-    }
-
-    // ========================================================================================
-    // Helpers
-    // ========================================================================================
-    private void waitFor(long milliseconds) {
-        try {
-            Thread.sleep(milliseconds);
-        } catch (InterruptedException e) {
-            fail("thread sleep error: " + e.getLocalizedMessage());
-        }
     }
 }
